ai/requirements.txt
# ai/requirements.txt
Flask
Flask-Cors
numpy
psycopg2-binary
scikit-learn
pandas
spacy